Add PocketCGB - Weatherly - Danny Boy (#21)
This is first downloadable content for pocketcdg, it is hard to find or make correct cdg combination. Original song is made by Frederick Weatherly in 1913. This work is in the public domain in the United States because it was published before January 1, 1926. The author died in 1929, so this work is also in the public domain in countries and areas where the copyright term is the author's life plus 80 years or less. This work may also be in the public domain in countries and areas with longer native copyright terms that apply the rule of the shorter term to foreign works. Check: https://en.wikisource.org/wiki/Danny_Boy